requirements.txt
streamlit
pandas
pytest
python-dotenv